refactor(daemon): extract transcript normalization
